KULRICE-5635 added test to verify document status and category doc search add test and fix to exceptionroutingserviceimpl to ensure placeInExceptionRouting works when invoked via the WorkflowDocument API

+++ impl/src/main/java/org/kuali/rice/kew/messaging/exceptionhandling/ExceptionRoutingServiceImpl.java
@@ -121,7 +121,12 @@ public class ExceptionRoutingServiceImpl implements WorkflowDocumentExceptionRou
             	throw new RiceRuntimeException("Failed to generate exception requests for exception routing!");
             }
             activateExceptionRequests(routeContext, exceptionRequests, errorMessage, invokePostProcessor);
-            KSBServiceLocator.getMessageQueueService().delete(persistedMessage);
+
+            if (persistedMessage == null) {
+                LOG.warn("Attempting to delete null persisted message.");
+            } else {
+                KSBServiceLocator.getMessageQueueService().delete(persistedMessage);
+            }
         } finally {
             performanceLogger.log("Time to generate exception request.");
             MDC.remove("docId");
